+++ vtm-android-example/src/org/oscim/android/test/Samples.java
@@ -53,6 +53,7 @@ public class Samples extends Activity {
         linearLayout.addView(createButton(VectorLayerMapActivity.class));
         linearLayout.addView(createButton(MultiMapActivity.class));
         linearLayout.addView(createButton(MapFragmentActivity.class));
+        linearLayout.addView(createButton(NewGesturesActivity.class));
     }
 
     private Button createButton(final Class<?> clazz) {