+++ vtm/src/org/oscim/layers/MapEventLayer2.java
@@ -22,11 +22,13 @@ import org.oscim.core.MapPosition;
 import org.oscim.core.Tile;
 import org.oscim.event.Event;
 import org.oscim.event.Gesture;
-import org.oscim.event.GestureListener;
 import org.oscim.event.MotionEvent;
 import org.oscim.map.Map;
 import org.oscim.map.Map.InputListener;
 import org.oscim.map.ViewController;
+
+import java.util.Timer;
+import java.util.TimerTask;
 
 import static org.oscim.backend.CanvasAdapter.dpi;
 import static org.oscim.utils.FastMath.withinSquaredDist;
@@ -37,7 +39,7 @@ import static org.oscim.utils.FastMath.withinSquaredDist;
  * TODO rewrite using gesture primitives to build more complex gestures:
  * maybe something similar to this https://github.com/ucbvislab/Proton
  */
-public class MapEventLayer extends AbstractMapEventLayer implements InputListener, GestureListener {
+public class MapEventLayer2 extends AbstractMapEventLayer implements InputListener {
 
     private boolean mEnableRotate = true;
     private boolean mEnableTilt = true;
@@ -56,8 +58,12 @@ public class MapEventLayer extends AbstractMapEventLayer implements InputListene
     private boolean mDoTilt;
 
     private boolean mDown;
-    private boolean mDoubleTap;
     private boolean mDragZoom;
+    private boolean mTwoFingers;
+    private boolean mTwoFingersDone;
+    private int mTaps;
+    private long mStartDown;
+    private MotionEvent mLastTap;
 
     private float mPrevX1;
     private float mPrevY1;
@@ -79,18 +85,30 @@ public class MapEventLayer extends AbstractMapEventLayer implements InputListene
     private static final float PINCH_ROTATE_THRESHOLD = 0.2f;
     private static final float PINCH_ROTATE_THRESHOLD2 = 0.5f;
 
+    //TODO Should be initialized with platform specific defaults
     /**
      * 100 ms since start of move to reduce fling scroll
      */
-    private static final float FLING_MIN_THREHSHOLD = 100;
+    private static final long FLING_MIN_THRESHOLD = 100;
+    private static final long DOUBLE_TAP_THRESHOLD = 300;
+    private static final long LONG_PRESS_THRESHOLD = 500;
 
     private final VelocityTracker mTracker;
+    private final Timer mTimer;
+    private TimerTask mTimerTask;
 
     private final MapPosition mapPosition = new MapPosition();
 
-    public MapEventLayer(Map map) {
+    public MapEventLayer2(Map map) {
         super(map);
         mTracker = new VelocityTracker();
+        mTimer = new Timer();
+    }
+
+    @Override
+    public void onDetach() {
+        mTimer.cancel();
+        mTimer.purge();
     }
 
     @Override
@@ -146,24 +164,58 @@ public class MapEventLayer extends AbstractMapEventLayer implements InputListene
         mFixOnCenter = enable;
     }
 
-    boolean onTouchEvent(MotionEvent e) {
-
+    private boolean onTouchEvent(final MotionEvent e) {
         int action = getAction(e);
+        final long time = e.getTime();
 
         if (action == MotionEvent.ACTION_DOWN) {
-            mMap.animator().cancel();
+            if (mTimerTask != null) {
+                mTimerTask.cancel();
+                mTimer.purge();
+                mTimerTask = null;
+            }
+            mMap.handleGesture(Gesture.PRESS, e);
+            mDown = true;
+            mStartDown = time;
+            if (mTaps > 0) {
+                float mx = e.getX(0) - mLastTap.getX();
+                float my = e.getY(0) - mLastTap.getY();
+                if (isMinimalMove(mx, my)) {
+                    mTaps = 0;
+                    //log.debug("tap {} {}", mLastTap.getX(), mLastTap.getY());
+                    mMap.handleGesture(Gesture.TAP, mLastTap);
+                }
+            } else {
+                mMap.animator().cancel();
 
-            mStartMove = -1;
-            mDoubleTap = false;
-            mDragZoom = false;
+                mStartMove = -1;
+                mDragZoom = false;
+                mTwoFingers = false;
+                mTwoFingersDone = false;
+
+                mTimerTask = new TimerTask() {
+                    @Override
+                    public void run() {
+                        if (mTwoFingers || mStartMove != -1)
+                            return;
+                        mMap.post(new Runnable() {
+                            @Override
+                            public void run() {
+                                //log.debug("long press {} {}", e.getX(), e.getY());
+                                mMap.handleGesture(Gesture.LONG_PRESS, e);
+                            }
+                        });
+                    }
+                };
+                mTimer.schedule(mTimerTask, LONG_PRESS_THRESHOLD);
+            }
 
             mPrevX1 = e.getX(0);
             mPrevY1 = e.getY(0);
 
-            mDown = true;
             return true;
         }
-        if (!(mDown || mDoubleTap)) {
+        if (!mDown) {
             /* no down event received */
             return false;
         }
@@ -174,17 +226,12 @@ public class MapEventLayer extends AbstractMapEventLayer implements InputListene
         }
         if (action == MotionEvent.ACTION_UP) {
             mDown = false;
-            if (mDoubleTap && !mDragZoom) {
-                float pivotX = 0, pivotY = 0;
-                if (!mFixOnCenter) {
-                    pivotX = mPrevX1 - mMap.getWidth() / 2;
-                    pivotY = mPrevY1 - mMap.getHeight() / 2;
-                }
-
-                /* handle double tap zoom */
-                mMap.animator().animateZoom(300, 2, pivotX, pivotY);
-
-            } else if (mStartMove > 0) {
+            if (mTimerTask != null) {
+                mTimerTask.cancel();
+                mTimer.purge();
+                mTimerTask = null;
+            }
+            if (mStartMove > 0) {
                 /* handle fling gesture */
                 mTracker.update(e.getX(), e.getY(), e.getTime());
                 float vx = mTracker.getVelocityX();
@@ -192,16 +239,86 @@ public class MapEventLayer extends AbstractMapEventLayer implements InputListene
 
                 /* reduce velocity for short moves */
                 float t = e.getTime() - mStartMove;
-                if (t < FLING_MIN_THREHSHOLD) {
-                    t = t / FLING_MIN_THREHSHOLD;
+                if (t < FLING_MIN_THRESHOLD) {
+                    t = t / FLING_MIN_THRESHOLD;
                     vy *= t * t;
                     vx *= t * t;
                 }
                 doFling(vx, vy);
             }
+
+            if (time - mStartDown > LONG_PRESS_THRESHOLD) {
+                //log.debug("  not a tap");
+                // this was not a tap
+                mTaps = 0;
+                return true;
+            }
+
+            if (mTaps > 0) {
+                if ((time - mLastTap.getTime()) >= DOUBLE_TAP_THRESHOLD) {
+                    mTaps = 1;
+                    //log.debug("tap {} {}", mLastTap.getX(), mLastTap.getY());
+                    mMap.handleGesture(Gesture.TAP, mLastTap);
+                } else {
+                    mTaps += 1;
+                }
+            } else {
+                mTaps = 1;
+            }
+
+            if (mLastTap != null) {
+                mLastTap.recycle();
+            }
+            mLastTap = e.copy();
+
+            if (mTaps == 3) {
+                mTaps = 0;
+                //log.debug("triple tap {} {}", e.getX(), e.getY());
+                mMap.handleGesture(Gesture.TRIPLE_TAP, e);
+            } else if (mTaps == 2) {
+                mTimerTask = new TimerTask() {
+                    @Override
+                    public void run() {
+                        mTaps = 0;
+                        if (mDragZoom)
+                            return;
+                        mMap.post(new Runnable() {
+                            @Override
+                            public void run() {
+                                //log.debug("double tap {} {}", e.getX(), e.getY());
+                                if (!mMap.handleGesture(Gesture.DOUBLE_TAP, e)) {
+                                    /* handle double tap zoom */
+                                    final float pivotX = mFixOnCenter ? 0 : mPrevX1 - mMap.getWidth() / 2;
+                                    final float pivotY = mFixOnCenter ? 0 : mPrevY1 - mMap.getHeight() / 2;
+                                    mMap.animator().animateZoom(300, 2, pivotX, pivotY);
+                                }
+                            }
+                        });
+                    }
+                };
+                mTimer.schedule(mTimerTask, DOUBLE_TAP_THRESHOLD);
+            } else {
+                mTimerTask = new TimerTask() {
+                    @Override
+                    public void run() {
+                        mTaps = 0;
+                        if (!mTwoFingers && mStartMove == -1) {
+                            mMap.post(new Runnable() {
+                                @Override
+                                public void run() {
+                                    //log.debug("tap {} {}", e.getX(), e.getY());
+                                    mMap.handleGesture(Gesture.TAP, e);
+                                }
+                            });
+                        }
+                    }
+                };
+                mTimer.schedule(mTimerTask, DOUBLE_TAP_THRESHOLD);
+            }
             return true;
         }
         if (action == MotionEvent.ACTION_CANCEL) {
+            mTaps = 0;
             return false;
         }
         if (action == MotionEvent.ACTION_POINTER_DOWN) {
@@ -210,6 +327,12 @@ public class MapEventLayer extends AbstractMapEventLayer implements InputListene
             return true;
         }
         if (action == MotionEvent.ACTION_POINTER_UP) {
+            if (e.getPointerCount() == 2 && !mTwoFingersDone) {
+                //log.debug("two finger tap");
+                if (!mMap.handleGesture(Gesture.TWO_FINGER_TAP, e)) {
+                    mMap.animator().animateZoom(300, 0.5, 0f, 0f);
+                }
+            }
             updateMulti(e);
             return true;
         }
@@ -237,12 +360,7 @@ public class MapEventLayer extends AbstractMapEventLayer implements InputListene
             mPrevY1 = y1;
 
             /* double-tap drag zoom */
-            if (mDoubleTap) {
-                /* just ignore first move event to set mPrevX/Y */
-                if (!mDown) {
-                    mDown = true;
-                    return;
-                }
+            if (mTaps == 1) {
                 if (!mDragZoom && !isMinimalMove(mx, my)) {
                     mPrevX1 -= mx;
                     mPrevY1 -= my;
@@ -305,6 +423,7 @@ public class MapEventLayer extends AbstractMapEventLayer implements InputListene
                     mCanScale = false;
                     mCanRotate = false;
                     mDoTilt = true;
+                    mTwoFingersDone = true;
                 }
             }
         }
@@ -331,6 +450,7 @@ public class MapEventLayer extends AbstractMapEventLayer implements InputListene
                     /* start rotate, disable tilt */
                     mDoRotate = true;
                     mCanTilt = false;
+                    mTwoFingersDone = true;
 
                     mAngle = rad;
                 } else if (!mDoScale) {
@@ -350,6 +470,7 @@ public class MapEventLayer extends AbstractMapEventLayer implements InputListene
                 mDoRotate = true;
                 mCanRotate = true;
                 mAngle = rad;
+                mTwoFingersDone = true;
             }
         }
 
@@ -365,6 +486,7 @@ public class MapEventLayer extends AbstractMapEventLayer implements InputListene
 
                     mCanTilt = false;
                     mDoScale = true;
+                    mTwoFingersDone = true;
                 }
             }
             if (mDoScale || mDoRotate) {
@@ -413,6 +535,8 @@ public class MapEventLayer extends AbstractMapEventLayer implements InputListene
         mPrevY1 = e.getY(0);
 
         if (cnt == 2) {
+            mTwoFingers = true;
+
             mDoScale = false;
             mDoRotate = false;
             mDoTilt = false;
@@ -443,15 +567,6 @@ public class MapEventLayer extends AbstractMapEventLayer implements InputListene
         mMap.animator().animateFling(velocityX * 2, velocityY * 2,
                 -w, w, -h, h);
         return true;
-    }
-
-    @Override
-    public boolean onGesture(Gesture g, MotionEvent e) {
-        if (g == Gesture.DOUBLE_TAP) {
-            mDoubleTap = true;
-            return true;
-        }
-        return false;
     }
 
     private static class VelocityTracker {